% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_msa.R
\name{geom_msa}
\alias{geom_msa}
\title{geom_msa}
\usage{
geom_msa(mapping = NULL, data = NULL, ...)
}
\arguments{
\item{mapping}{aesthetic mapping}

\item{data}{data}

\item{...}{additional parameters passed to geom_tile}
}
\value{
ggplot msa layer
}
\description{
add msa layer
}
\author{
Bradley R Jones
}
