% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_seqlab.R
\name{geom_seqlab}
\alias{geom_seqlab}
\title{geom_msa}
\usage{
geom_seqlab(mapping = NULL, data = NULL, ...)
}
\arguments{
\item{mapping}{aesthetic mapping}

\item{data}{data}

\item{...}{additional parameters passed to geom_polygon}
}
\value{
ggplot sequence label layer
}
\description{
add sequence label layer
}
\author{
guangchuang yu, Bradley R Jones
}
