% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gghighlighter.R
\name{gghighlighter}
\alias{gghighlighter}
\title{gghighlighter}
\usage{
gghighlighter(fasta, start = NULL, end = NULL,
  palette = c("Chemistry_AA", "Shapely_AA", "Zappo_AA", "Taylor_AA",
  "LANL_AA", "Chemistry_Nucle", "Shapely_Nucle", "Zappo_Nucle",
  "Taylor_Nucle"), ...)
}
\arguments{
\item{fasta}{aligned fasta file}

\item{start}{start position to plot}

\item{end}{end position to plot}

\item{palette}{name of highlighter colour scheme (5 amino acid color schemes, 4 nucleic acid color schemes.)}

\item{...}{additional parameters passed to fortify and ggplot}

\item{consensus}{either the name of the consensus sequence or e a character string of the consensus sequence or NA for no consensus highlighting}

\item{highlight.diff}{TRUE to highlight differences from consesnsus and FALSE to highlight similarities  (default: TRUE)}

\item{remove.blank}{character vector of characters to mark as NA (default: NULL)}
}
\value{
ggplot object
}
\description{
plot multiple sequence alignment using ggplot2
}
\author{
Bradley R Jones, guangchuang yu
}
