% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msa_utils.R
\name{infer_character_logos}
\alias{infer_character_logos}
\title{infer_character_logos}
\usage{
infer_character_logos(y, font = "helvetica_regular", ...)
}
\arguments{
\item{y}{fortified alignment}

\item{font}{character font}

\item{...}{params passed by other functions}
}
\value{
tibble
}
\description{
add logos to fortified alignment
}
\author{
guangchuang yu, Bradley R Jones
}
