% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_highlighter.R
\name{scale_fill_highlighter}
\alias{scale_fill_highlighter}
\title{scale_fill_highlighter}
\usage{
scale_fill_highlighter(palette = c("Chemistry_AA", "Shapely_AA",
  "Zappo_AA", "Taylor_AA", "LANL_AA", "Chemistry_Nucle", "Shapely_Nucle",
  "Zappo_Nucle", "Taylor_Nucle"), ...)
}
\arguments{
\item{palette}{name of highlighter colour palette}

\item{...}{additional parameters passed to scale_fill_manual}
}
\value{
ggplot fill scale
}
\description{
Nucleotide and Amino acid colour scales
}
\author{
Bradley R Jones, guangchuang yu
}
