% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_msa.R
\name{scale_fill_msa}
\alias{scale_fill_msa}
\title{scale_fill_msa}
\usage{
scale_fill_msa(palette = NULL, ...)
}
\arguments{
\item{palette}{name of msa colour palette}

\item{...}{additional parameters passed to scale_fill_manual}
}
\value{
ggplot fill scale
}
\description{
Nucleotide and Amino acid colour scales
}
\author{
guangchuang yu, Bradley R Jones
}
