% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-fortify.R
\name{fortify_alignment_matrix}
\alias{fortify_alignment_matrix}
\title{fortify_alignment_matrix}
\usage{
fortify_alignment_matrix(x, start = NULL, end = NULL, consensus = NA,
  highlight.diff = TRUE, remove.blank = NULL, ...)
}
\arguments{
\item{x}{alignment matrix}

\item{start}{start position to plot}

\item{end}{end position to plot}

\item{consensus}{either the name of the consensus sequence or e a character string of the consensus sequence or NA for no consensus highlighting}

\item{highlight.diff}{TRUE to highlight differences from consesnsus and FALSE to highlight similarities (default: TRUE)}

\item{remove.blank}{character vector of characters to mark as NA (default: NULL)}

\item{...}{additional parameters passed to geom_tile}
}
\value{
tibble
}
\description{
fortify alignment matrix
}
\author{
Bradley R Jones
}
