% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-fortify.R
\name{fortify_alignment_matrix}
\alias{fortify_alignment_matrix}
\title{fortify_alignment_matrix}
\usage{
fortify_alignment_matrix(x, consensus = NA, remove.blank = c("X", "x",
  "-", "?", "*"), ...)
}
\arguments{
\item{x}{alignment matrix}

\item{consensus}{either the name of the consensus sequence or the a character string of the consensus sequence or NA for no ceonsenus highlighting}

\item{remove.blank}{character vector of characters to mark as NA or NULL (default: "X", "x", "-", "?", "*")}

\item{...}{additional parameters passed to geom_tile}
}
\value{
tibble
}
\description{
fortify alignment matrix
}
\author{
Bradley R Jones
}
