% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmsa.R
\name{ggmsa}
\alias{ggmsa}
\title{ggmsa}
\usage{
ggmsa(fasta, consensus = NA, fill.scale = ggplot2::scale_fill_discrete)
}
\arguments{
\item{fasta}{aligned fasta file}

\item{consensus}{either the name of the consensus sequence or the a character string of the consensus sequence or NA for no ceonsenus highlighting}

\item{fill.scale}{fill scale}
}
\value{
ggplot object
}
\description{
plot multiple sequence alignment using ggplot2
}
\author{
guangchuang yu, Bradley R Jones
}
