% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmsa.R
\name{ggmsa}
\alias{ggmsa}
\title{ggmsa}
\usage{
ggmsa(fasta, start = NULL, end = NULL, font = "helvetica_regular",
  color = c("Chemistry_AA", "Shapely_AA", "Zappo_AA", "Taylor_AA",
  "LANL_AA", "Chemistry_Nucle", "Shapely_Nucle", "Zappo_Nucle",
  "Taylor_Nucle"), ...)
}
\arguments{
\item{fasta}{aligned fasta file}

\item{start}{start position to plot}

\item{end}{end position to plot}

\item{font}{character font}

\item{color}{name of msa colour scheme (6 amino acid color schemes, 4 nucleic acid color schemes. Note: Culstal is an amino acid color scheme)}

\item{...}{additional parameters passed to fortify and ggplot}

\item{consensus}{either the name of the consensus sequence or e a character string of the consensus sequence or NA for no consensus highlighting}

\item{remove.blank}{character vector of characters to mark as NA (default: NULL)}
}
\value{
ggplot object
}
\description{
plot multiple sequence alignment using ggplot2
}
\author{
guangchuang yu
}
