% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmsa.R
\name{ggmsa}
\alias{ggmsa}
\title{ggmsa}
\usage{
ggmsa(fasta, start = NULL, end = NULL, font = "helvetica_regular",
  ..., palette = NULL)
}
\arguments{
\item{fasta}{aligned fasta file}

\item{start}{start position to plot}

\item{end}{end position to plot}

\item{font}{character font}

\item{...}{additional parameters passed to fortify and ggplot}

\item{palette}{name of msa colour palette}

\item{consensus}{either the name of the consensus sequence or e a character string of the consensus sequence or NA for no consensus highlighting}

\item{remove.blank}{character vector of characters to mark as NA (default: NULL)}
}
\value{
ggplot object
}
\description{
plot multiple sequence alignment using ggplot2
}
\author{
guangchuang yu, Bradley R Jones
}
